\name{BVR2XYZ}
\alias{BVR2XYZ}
\title{convert from BVR coordinates to XYZ
temperature (Robertson)}
\description{\code{BVR2XYZ}  convert from BVR coordinates to XYZ.
}
\usage{BVR2XYZ(BVRmatrix) }
\arguments{
  \item{BVRmatrix}{ BVR coordinates}
}
\value{
  XYZ coordinates
}
\source{
Bruce Justin Lindbloom, 2014
\url{http://www.brucelindbloom.com/index.html?Eqn_XYZ_to_T.html}
}
\references{
Bruce Justin Lindbloom, 2014
\url{http://www.brucelindbloom.com/index.html?Eqn_XYZ_to_T.html}
}
\author{Jose Gama}
\keyword{datasets}

