\name{cccie31}
\alias{cccie31}
\title{CIE (1931) 2-deg chromaticity coordinates}
\usage{cccie31}
\description{\code{cccie31} is a table with CIE (1931) 2-deg 
chromaticity coordinates.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{x}{ x chromaticity coordinate}
    \item{y}{ y chromaticity coordinate}
    \item{z}{ z chromaticity coordinate}
  }
}
\source{
Commission Internationale de l'Eclairage Proceedings, 1931
Cambridge: Cambridge University Press.
}
\references{
Commission Internationale de l'Eclairage Proceedings, 1931
Cambridge: Cambridge University Press.
}
\author{Jose Gama}
\examples{
data(cccie31)
cccie31
}
\keyword{datasets}

