\name{cccie64}
\alias{cccie64}
\title{CIE (1964) 10-deg chromaticity coordinates}
\usage{cccie64}
\description{\code{cccie64} is a table with CIE (1964) 10-deg 
chromaticity coordinates.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{x}{ x chromaticity coordinate}
    \item{y}{ y chromaticity coordinate}
    \item{z}{ z chromaticity coordinate}
  }
}
\source{
Wyszecki, G., & Stiles, W. S., 1982
Color Science: concepts and methods, quantitative data and formulae
(2nd ed.). New York: Wiley.
}
\references{
Wyszecki, G., & Stiles, W. S., 1982
Color Science: concepts and methods, quantitative data and formulae
(2nd ed.). New York: Wiley.
}
\author{Jose Gama}
\examples{
data(cccie64)
cccie64
}
\keyword{datasets}

