\name{CCT2XYZ}
\alias{CCT2XYZ}
\title{Convert CCT to XYZ}
\description{\code{CCT2XYZ} Converts correlated color temperature (CCT) to 
CIE tristimulus XYZ.
}
\usage{CCT2XYZ(CCTmatrix) }
\arguments{
  \item{CCTmatrix}{ CCT values}
}
\value{
  CIE tristimulus XYZ
}
\source{
Bruce Justin Lindbloom, 2013 
Color Calculator
\url{www.brucelindbloom.com}
}
\references{
Bruce Justin Lindbloom, 2013 
Color Calculator
\url{www.brucelindbloom.com}
}
\author{Jose Gama}
\examples{
CCT2XYZ(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

