\name{CentralsISCCNBS}
\alias{CentralsISCCNBS}
\title{Central notations for the revised ISCC-NBS}
\usage{CentralsISCCNBS}
\description{\code{CentralsISCCNBS} is a table with Central notations
for the revised ISCC-NBS Color-Name Blocks.}
\format{
  This data frame contains the following data:
  \describe{
    \item{Number}{ notation number}
    \item{Name}{ notation name}
    \item{MunsellSpec}{ Munsell specification}
  }
}
\references{
Kelly, Kenneth Low, 1910
Central notations for the revised ISCC-NBS Color-Name Blocks 
Journal of Research of the National Bureau of Standards
Research Paper 2911, Vol. 61  No. 5, November 1958
}
\author{Glenn Davis}
\examples{
data(CentralsISCCNBS)
CentralsISCCNBS
}
\keyword{datasets}

