\name{CheckColorLookup}
\alias{CheckColorLookup}
\title{convert Munsell notation to numeric HVC}
\description{\code{CheckColorLookup} Converts convert Munsell notation 
to numeric Hue,Value,Chroma.
}
\usage{CheckColorLookup(DataISCCNBS) }
\arguments{
  \item{DataISCCNBS}{ data.frame with columns MunsellSpec and Number}
}
\value{
  numeric Hue,Value,Chroma
}
\author{Glenn Davis}
\keyword{datasets}

