\name{CIE1976uv2CIE1931xy}
\alias{CIE1976uv2CIE1931xy}
\title{CIE-1976  u'v' to CIE-1931 xy}
\description{\code{CIE1976uv2CIE1931xy} CIE-1976  u'v' to CIE-1931 xy.
}
\usage{CIE1976uv2CIE1931xy(uvmatrix)}
\arguments{
    \item{uvmatrix}{ CIE-1976  u'v' data}
}
\source{
Paul Schils, 2014
Color theory phenomena
\url{http://www.color-theory-phenomena.nl/10.03.htm}
}
\references{
Paul Schils, 2014
Color theory phenomena
\url{http://www.color-theory-phenomena.nl/10.03.htm}
}
\author{Jose Gama}
\examples{
CIE1976uv2CIE1931xy(c(0.2830965, 0.4661789))
}
\keyword{datasets}

