\name{CIE1976uv2CIE1960uv}
\alias{CIE1976uv2CIE1960uv}
\title{CIE-1976 u'v' to  CIE-1960 uv}
\description{\code{CIE1976uv2CIE1960uv} CIE-1976 u'v' to  CIE-1960 uv.
}
\usage{CIE1976uv2CIE1960uv(uvmatrix)}
\arguments{
    \item{uvmatrix}{ CIE-1976  u'v' data}
}
\source{
Paul Schils, 2014
Color theory phenomena
\url{http://www.color-theory-phenomena.nl/10.03.htm}
}
\references{
Paul Schils, 2014
Color theory phenomena
\url{http://www.color-theory-phenomena.nl/10.03.htm}
}
\author{Jose Gama}
\examples{
CIE1976uv2CIE1960uv(c(0.2830965, 0.4661789))
}
\keyword{datasets}

