\name{CMY2CMYK}
\alias{CMY2CMYK}
\title{Convert CMY coordinates to CMYK}
\description{\code{CMY2CMYK} Converts CMY coordinates to CMYK.
}
\usage{CMY2CMYK(CMYmatrix) }
\arguments{
  \item{CMYmatrix}{ CMY coordinates}
}
\value{
  CMYK coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
CMY2CMYK(c(0.59072, 0.85570, 0.80283))
}
\keyword{datasets}

