\name{CMY2RGB}
\alias{CMY2RGB}
\title{Convert CMYK coordinates to RGB}
\description{\code{CMY2RGB} Converts CMYK coordinates to RGB.
}
\usage{CMY2RGB(CMYmatrix) }
\arguments{
  \item{CMYmatrix}{ CMY coordinates}
}
\value{
  RGB coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
CMY2RGB(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

