\name{CMYK2CMY}
\alias{CMYK2CMY}
\title{Convert CMYK coordinates to CMY}
\description{\code{CMYK2CMY} Converts CMYK coordinates to CMY.
}
\usage{CMYK2CMY(CMYKmatrix) }
\arguments{
  \item{CMYKmatrix}{ CMYK coordinates}
}
\value{
  CMY coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
CMYK2CMY(c(.342, .768, .683, .378))
}
\keyword{datasets}

