\name{conversionIlluminance}
\alias{conversionIlluminance}
\title{Conversion Factors for Units of Illuminance}
\usage{conversionIlluminance}
\description{\code{conversionIlluminance} is a table of conversion factors for units of Illuminance}
\format{
  This data frame contains the following columns:
  \describe{
    \item{footcandles}{ foot-candles}
    \item{lux}{ lm/m2 = lux}
    \item{phot}{ phot}
    \item{milliphot}{ milliphot}
    \item{units}{ units}
  }
}
\source{
J. Meyer-Arendt, "Radiometry and Photometry: Units and Conversion Factors," Appl. Opt.  7, 2081-2081 (1968).
}
\references{
J. Meyer-Arendt, "Radiometry and Photometry: Units and Conversion Factors," Appl. Opt.  7, 2081-2081 (1968).
}
\author{Jose Gama}
\examples{
data(conversionIlluminance)
conversionIlluminance
}
\keyword{datasets}

