\name{DeMarcoPokornySmith2degConeFundamentals1992}
\alias{DeMarcoPokornySmith2degConeFundamentals1992}
\title{DeMarco, Pokorny & Smith (1992) versions of the Smith-Pokorny 2-deg fundamentals}
\usage{DeMarcoPokornySmith2degConeFundamentals1992}
\description{\code{DeMarcoPokornySmith2degConeFundamentals1992} 
DeMarco, Pokorny & Smith (1992) versions of the Smith-Pokorny 
2-deg fundamentals based on the CIE Judd-Vos 2-deg CMFs.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{L2}{ L-cone spectral sensitivity, L2(lambda)}
    \item{M2}{ M-cone spectral sensitivity, M2(lambda)}
    \item{S2}{ S-cone spectral sensitivity,  S2(lambda)}
  }
}
\source{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\references{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\author{Jose Gama}
\examples{
data(DeMarcoPokornySmith2degConeFundamentals1992)
DeMarcoPokornySmith2degConeFundamentals1992
}
\keyword{datasets}

