\name{HSV2RGB}
\alias{HSV2RGB}
\title{Convert HSV coordinates to RGB}
\description{\code{HSV2RGB} Converts HSV coordinates to RGB.
}
\usage{HSV2RGB(HSVmatrix) }
\arguments{
  \item{HSVmatrix}{ HSV coordinates}
}
\value{
  RGB coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
HSV<-c(0.9590164, 0.4919355, 0.4862745)
HSV2RGB(HSV)
}
\keyword{datasets}

