\name{Hue.2.RGB}
\alias{Hue.2.RGB}
\title{Convert Hue to RGB}
\description{\code{Hue.2.RGB} Converts Hue to RGB for HSL conversion.
}
\usage{Hue.2.RGB(v1, v2, vH) }
\arguments{
  \item{v1}{ value 1}
  \item{v2}{ value 2}
  \item{vH}{ value of hue}
}
\value{
  RGB coordinates
}
\source{
Paul Centore 2014
The Munsell and Kubelka-Munk Toolbox
\url{http://www.99main.com/~centore/MunsellAndKubelkaMunkToolbox/MunsellAndKubelkaMunkToolbox.html}
}
\references{
Paul Centore 2014
The Munsell and Kubelka-Munk Toolbox
\url{http://www.99main.com/~centore/MunsellAndKubelkaMunkToolbox/MunsellAndKubelkaMunkToolbox.html}
}
\author{Jose Gama}
\examples{
Hue.2.RGB(1,2,3)
}
\keyword{datasets}

