\name{huedegree}
\alias{huedegree}
\title{convert Munsell hue to degree}
\description{\code{huedegree} convert Munsell hue to degree.
}
\usage{huedegree(MunIn) }
\arguments{
  \item{MunIn}{ Munsell hue color}
}
\value{
  Munsell hue degree
}
\source{
Takahiro Onodera, 2010
Color-Model-Munsell-Util 
\url{http://annocpan.org/dist/Color-Model-Munsell-Util}
}
\references{
Takahiro Onodera, 2010
Color-Model-Munsell-Util 
\url{http://annocpan.org/dist/Color-Model-Munsell-Util}
}
\author{Jose Gama}
\examples{
huedegree('1P')
}
\keyword{datasets}

