\name{illuminantD65}
\alias{illuminantD65}
\title{Relative spectral power distributions of CIE illuminant D65 at 1 
nm interval}
\usage{illuminantD65}
\description{\code{illuminantD65} is a table with Relative spectral power 
distributions of CIE illuminant D65 at 1 nm interval.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{intensity}{ Relative spectral power}
  }
}
\source{
Wyszecki, G., & Stiles, W. S., 1982
Color Science: concepts and methods, quantitative data and formulae
(2nd ed.). New York: Wiley.
}
\references{
Wyszecki, G., & Stiles, W. S., 1982
Color Science: concepts and methods, quantitative data and formulae
(2nd ed.). New York: Wiley.
}
\author{Jose Gama}
\examples{
data(illuminantD65)
illuminantD65
}
\keyword{datasets}

