\name{illuminants}
\alias{illuminants}
\title{Relative spectral power distributions of CIE illuminants at 5 
nm interval}
\usage{illuminants}
\description{\code{illuminants} is a table with Relative spectral power 
distributions of CIE illuminants at 5 nm interval.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{A}{ illuminant A}
    \item{B}{ illuminant B}
    \item{C}{ illuminant C}
    \item{D50}{ illuminant D50}
    \item{D55}{ illuminant D55}
    \item{D65}{ illuminant D65}
    \item{D75}{ illuminant D75}
    \item{D93}{ illuminant D93}
    \item{E}{ illuminant E}
    \item{Natural}{ illuminant Natural}
    \item{PlusWhite}{ illuminant PlusWhite}
    \item{TL84}{ illuminant TL84}
    \item{Polylux3000}{ illuminant Polylux3000}
    \item{Polylux4000}{ illuminant Polylux4000}
    \item{KolorRite}{ illuminant KolorRite}
    \item{FL1}{ illuminant FL1}
    \item{FL2}{ illuminant FL2}
    \item{FL3}{ illuminant FL3}
    \item{FL4}{ illuminant FL4}
    \item{FL5}{ illuminant FL5}
    \item{FL6}{ illuminant FL6}
    \item{FL7}{ illuminant FL7}
    \item{FL8}{ illuminant FL8}
    \item{FL9}{ illuminant FL9}
    \item{FL10}{ illuminant FL10}
    \item{FL11}{ illuminant FL11}
    \item{FL12}{ illuminant FL12}
  }
}
\source{
Wyszecki, G., & Stiles, W. S., 1982
Color Science: concepts and methods, quantitative data and formulae
(2nd ed.). New York: Wiley.
}
\references{
Wyszecki, G., & Stiles, W. S., 1982
Color Science: concepts and methods, quantitative data and formulae
(2nd ed.). New York: Wiley.
}
\author{Jose Gama}
\examples{
data(illuminants)
illuminants
}
\keyword{datasets}

