\name{ISObrightnessReflectometerRSD}
\alias{ISObrightnessReflectometerRSD}
\title{Weighting factors for the calculation of ISO brightness
}
\usage{ISObrightnessReflectometerRSD}
\description{\code{ISObrightnessReflectometerRSD} is a table with the weighting 
factors for the calculation of ISO brightness.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wln}{ wavelength}
    \item{F}{ factor}
    \item{weights}{ weight}
  }
}
\source{
Scandinavian Pulp, paper and board, 2003
Basic equations for optical properties
SCAN-G 5:03 Revised 2003
}
\references{
Scandinavian Pulp, paper and board, 2003
Basic equations for optical properties
SCAN-G 5:03 Revised 2003
}
\author{Jose Gama}
\examples{
data(ISObrightnessReflectometerRSD)
ISObrightnessReflectometerRSD
}
\keyword{datasets}

