\name{Lab2LCHab}
\alias{Lab2LCHab}
\title{Convert CIE Lab coordinates to LCHab}
\description{\code{Lab2LCHab} Converts CIE Lab coordinates to LCHab.
}
\usage{Lab2LCHab(LabMatrix) }
\arguments{
  \item{LabMatrix}{ CIE Lab coordinates}
}
\value{
  LCHab coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
Lab2LCHab(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

