\name{LAB2LUV}
\alias{LAB2LUV}
\title{Convert CIE Lab coordinates to CIE Luv}
\description{\code{LAB2LUV} Converts CIE Lab coordinates to CIE Luv.
}
\usage{LAB2LUV(Labmatrix) }
\arguments{
  \item{Labmatrix}{ CIE Lab coordinates}
}
\value{
  XYZ coordinates
}
\author{Jose Gama}
\examples{
LAB2LUV(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

