\name{LCHab2Lab}
\alias{LCHab2Lab}
\title{Convert LCHab coordinates to CIE Lab}
\description{\code{LCHab2Lab} Converts LCHab coordinates to CIE Lab.
}
\usage{LCHab2Lab(LCHabmatrix) }
\arguments{
  \item{LCHabmatrix}{ LCHab coordinates}
}
\value{
  CIE Lab coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
LCHab2Lab(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

