\name{LCHuv2Luv}
\alias{LCHuv2Luv}
\title{Convert LCHuv coordinates to CIE Luv}
\description{\code{LCHuv2Luv} Converts LCHuv coordinates to CIE Luv.
}
\usage{LCHuv2Luv(LCHuvmatrix) }
\arguments{
  \item{LCHuvmatrix}{ LCHuv coordinates}
}
\value{
  CIE Luv coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
LCHuv2Luv(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

