\name{LEF2RGB}
\alias{LEF2RGB}
\title{Convert LEF coordinates to RGB}
\description{\code{LEF2RGB} Converts LEF coordinates to RGB.
}
\usage{LEF2RGB(LEFmatrix) }
\arguments{
  \item{LEFmatrix}{ LEF coordinates}
}
\value{
  RGB coordinates
}
\source{
Kang, Henry R, 2006
Computational color technology
Spie Press Bellingham
}
\references{
Kang, Henry R, 2006
Computational color technology
Spie Press Bellingham
}
\author{Jose Gama}
\examples{
LEF<-c(176.66667, 53.50000, -12.99038)
LEF2RGB(LEF)
}
\keyword{datasets}

