\name{LUV2LAB}
\alias{LUV2LAB}
\title{Convert CIE Luv coordinates to CIE Lab}
\description{\code{LUV2LAB} Converts CIE Luv coordinates to CIE Lab.
}
\usage{LUV2LAB(Luvmatrix) }
\arguments{
  \item{Luvmatrix}{ Luv matrix}
}
\value{
  XYZ coordinates
}
\author{Jose Gama}
\examples{
LUV2LAB(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

