\name{Luv2LCHuv}
\alias{Luv2LCHuv}
\title{Convert CIE Luv coordinates to LCHuv}
\description{\code{Luv2LCHuv} Converts CIE Luv coordinates to LCHuv.
}
\usage{Luv2LCHuv(LuvMatrix) }
\arguments{
  \item{LuvMatrix}{ Luv coordinates}
}
\value{
  LCHuv coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
Luv2LCHuv(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

