\name{MaterialReferenceData}
\alias{MaterialReferenceData}
\title{Material Reference Data from Principles of Digital Image Synthesis}
\usage{MaterialReferenceData}
\description{\code{MaterialReferenceData} Material Reference Data from 
Principles of Digital Image Synthesis, Appendix G
Andrew S. Glassner
16 August 1994.}
\format{
  This data frame contains the following data:
  \describe{
     \item{ wavelength }{ wavelength (nm)}
     \item{ DarkSkin }{ spectra of:   dark skin }
     \item{ LightSkin }{ spectra of:   light skin }
     \item{ BlueSky }{ spectra of:   blue sky }
     \item{ Foliage }{ spectra of:   foliage }
     \item{ BlueFlower }{ spectra of:   blue flower }
     \item{ BluishGreen }{ spectra of:   bluish green }
     \item{ Orange }{ spectra of:   orange }
     \item{ PurplishBlue }{ spectra of:   purplish blue }
     \item{ ModerateRed }{ spectra of:   moderate red }
     \item{ Purple }{ spectra of:   purple }
     \item{ YellowGreen }{ spectra of:   yellow green }
     \item{ OrangeYellow }{ spectra of:   orange yellow }
     \item{ Blue }{ spectra of:   blue }
     \item{ Green }{ spectra of:   green }
     \item{ Red }{ spectra of:   red }
     \item{ Yellow }{ spectra of:   yellow }
     \item{ Magenta }{ spectra of:   magenta }
     \item{ Cyan }{ spectra of:   cyan }
     \item{ White }{ spectra of:   white }
     \item{ Neutral }{ spectra of:   neutral }
     \item{ Neutral6.5 }{ spectra of:   neutral 6.5 }
     \item{ Neutral5 }{ spectra of:   neutral 5 }
     \item{ Neutral3.5 }{ spectra of:   neutral 3.5 }
     \item{ Black }{ spectra of:   black }
     \item{ PineNeedles }{ spectra of:   pine needles }
     \item{ SilverMapleLeaf }{ spectra of:   silver maple leaf }
     \item{ DarkGreenMapleLeaf }{ spectra of:   dark green maple leaf }
     \item{ RedMapleLeaf }{ spectra of:   red maple leaf }
     \item{ Grass }{ spectra of:   grass }
     \item{ Soil }{ spectra of:   soil }
     \item{ VineLeaf }{ spectra of:   vine leaf }
     \item{ Alphalt }{ spectra of:   alphalt }
     \item{ DaisyWhitePetals }{ spectra of:   daisy white petals }
     \item{ DaisyYellowCenter }{ spectra of:   daisy yellow center }
     \item{ MarigoldOrange }{ spectra of:   marigold orange }
     \item{ MarigoldYellow }{ spectra of:   marigold yellow }
     \item{ DarkBlueJeans }{ spectra of:   dark blue jeans }
     \item{ FadedJeans }{ spectra of:   faded jeans }
     \item{ DarkBlueSweatPants }{ spectra of:   dark blue sweat pants }
     \item{ Denim }{ spectra of:   denim }
     \item{ WheatBread }{ spectra of:   wheat bread }
     \item{ WheatBreadCrust }{ spectra of:   wheat bread crust }
     \item{ Pancake }{ spectra of:   pancake }
     \item{ SwissArmyKnife }{ spectra of:   swiss army knife }
     \item{ PineWood }{ spectra of:   pine wood }
     \item{ MapleWood }{ spectra of:   maple wood }
     \item{ OakWood }{ spectra of:   oak wood }
     \item{ Bamboo }{ spectra of:   bamboo }
     \item{ Redwood }{ spectra of:   redwood }
     \item{ WalnutWood }{ spectra of:   walnut wood }
     \item{ YellowBanana }{ spectra of:   yellow banana }
     \item{ RipeBrownBanana }{ spectra of:   ripe brown banana }
     \item{ Cucumber }{ spectra of:   cucumber }
     \item{ CornKernel }{ spectra of:   corn kernel }
     \item{ CornHusk }{ spectra of:   corn husk }
     \item{ YellowDeliciousApple }{ spectra of:   yellow delicious apple }
     \item{ GreenPepper }{ spectra of:   green pepper }
     \item{ LemonSkin }{ spectra of:   lemon skin }
     \item{ Lettuce }{ spectra of:   lettuce }
     \item{ Carrot }{ spectra of:   carrot }
     \item{ BarleySeeds }{ spectra of:   barley seeds }
     \item{ LentilSeeds }{ spectra of:   lentil seeds }
     \item{ BrownRiceSeeds }{ spectra of:   brown rice seeds }
     \item{ Sand }{ spectra of:   sand }
  }
}
\source{
Andrew S. Glassner, 1995
Principles of Digital Image Synthesis
The Morgan Kaufmann Series in Computer Graphics and Geometric Modeling
}
\references{
Andrew S. Glassner, 1995
Principles of Digital Image Synthesis
The Morgan Kaufmann Series in Computer Graphics and Geometric Modeling
}
\author{Jose Gama}
\examples{
data(MaterialReferenceData)
str(MaterialReferenceData)
}
\keyword{datasets}

