\name{Munsell100hues55}
\alias{Munsell100hues55}
\title{Chromaticity diagram showing values for x and y for Illuminant A
for 100 hues at 5/5
}
\usage{Munsell100hues55}
\description{\code{Munsell100hues55} Chromaticity diagram showing Tristimulus 
Values and Trilinear Coordinates for Illuminant A for 100 hues at 5/5.}
\format{
  This data frame contains the following columns:
  \describe{
    \item{BookNotation}{ Munsell color notation from the Munsell book}
    \item{MunsellProductionNo}{ }
    \item{X}{ Tristimulus Value X}
    \item{Y}{ Tristimulus Value Y}
    \item{Z}{ Tristimulus Value Z}
    \item{x}{ Trilinear Coordinate x}
    \item{y}{ Trilinear Coordinate y}
    \item{z}{ Trilinear Coordinate z}
  }
}
\source{
Hermann VON Schelling, Dean Farnsworth, 1949
Trichromatic Specifications of the Munsell 100 Hues at 5/5 for Illuminant a
Defense Technical Information Center
NAVAL MEDICAL RESEARCH LAB NEW LONDON CONN. 
}
\references{
Hermann VON Schelling, Dean Farnsworth, 1949
Trichromatic Specifications of the Munsell 100 Hues at 5/5 for Illuminant a
Defense Technical Information Center
NAVAL MEDICAL RESEARCH LAB NEW LONDON CONN. 
}
\author{Jose Gama}
\examples{
data(Munsell100hues55)
Munsell100hues55
}
\keyword{datasets}

