\name{Munsell100hues55FM100}
\alias{Munsell100hues55FM100}
\title{Munsell 100-Hues at 5/5, production numbers 101 to 200 and 
Farnsworth-Munsell 100 Hue test}
\usage{Munsell100hues55FM100}
\description{\code{Munsell100hues55FM100} Munsell 100-Hues at 5/5, 
production numbers 101 to 200 and Farnsworth-Munsell 100 Hue test.}
\format{
  This data frame contains the following columns:
  \describe{
    \item{FMtest}{ Farnsworth-Munsell 100 Hue test value}
    \item{MunsellNumber}{ Munsell 100-Hues at 5/5, production number}
  }
}
\source{
Hermann VON Schelling, Dean Farnsworth, 1949
Trichromatic Specifications of the Munsell 100 Hues at 5/5 for Illuminant a
Defense Technical Information Center
NAVAL MEDICAL RESEARCH LAB NEW LONDON CONN. 
}
\references{
Hermann VON Schelling, Dean Farnsworth, 1949
Trichromatic Specifications of the Munsell 100 Hues at 5/5 for Illuminant a
Defense Technical Information Center
NAVAL MEDICAL RESEARCH LAB NEW LONDON CONN. 
}
\author{Jose Gama}
\examples{
data(Munsell100hues55FM100)
Munsell100hues55FM100
}
\keyword{datasets}

