\name{MunsellNeutrals2sRGB}
\alias{MunsellNeutrals2sRGB}
\title{Table with Munsell Neutrals and corresponding sRGB}
\usage{MunsellNeutrals2sRGB}
\description{\code{MunsellNeutrals2sRGB} table with Munsell Neutrals 
and corresponding sRGB.}
\format{
  This data frame contains the following columns:
  \describe{
    \item{MunsellNeutral}{ Munsell N}
    \item{R}{ sRGB R}
    \item{G}{ sRGB G}
    \item{B}{ sRGB B}
  }
}
\source{
Paul Centore 2014
The Munsell and Kubelka-Munk Toolbox
\url{http://www.99main.com/~centore/MunsellAndKubelkaMunkToolbox/MunsellAndKubelkaMunkToolbox.html}

James D. Foley, Andries van Dam, Steven K. Feiner, & John F. Hughes, 1990
Computer Graphics: Principles and Practice, 2nd ed.,
Addison-Wesley Publishing Company.

Gunter Wyszecki & W. S. Stiles, 1982
Color Science: Concepts and Methods, Quantitative Data and Formulae, 2nd edition, John Wiley and Sons
}
\references{
Paul Centore 2014
The Munsell and Kubelka-Munk Toolbox
\url{http://www.99main.com/~centore/MunsellAndKubelkaMunkToolbox/MunsellAndKubelkaMunkToolbox.html}

James D. Foley, Andries van Dam, Steven K. Feiner, & John F. Hughes, 1990
Computer Graphics: Principles and Practice, 2nd ed.,
Addison-Wesley Publishing Company.

Gunter Wyszecki & W. S. Stiles, 1982
Color Science: Concepts and Methods, Quantitative Data and Formulae, 2nd edition, John Wiley and Sons
}
\author{Jose Gama}
\examples{
data(MunsellNeutrals2sRGB)
MunsellNeutrals2sRGB
}
\keyword{datasets}

