\name{MunsellSpecToHVC}
\alias{MunsellSpecToHVC}
\title{convert Munsell notation to numeric HVC}
\description{\code{MunsellSpecToHVC} Converts convert Munsell notation 
to numeric Hue,Value,Chroma.
}
\usage{MunsellSpecToHVC(MunsellSpecString) }
\arguments{
  \item{MunsellSpecString}{ string with the Munsell hue, value and chroma.}
}
\value{
  numeric Hue,Value,Chroma
}
\author{Glenn Davis}
\keyword{datasets}

