\name{MunsellSpectral}
\alias{MunsellSpectral}
\title{Table with Munsell spectral data}
\usage{MunsellSpectral}
\description{\code{MunsellSpectral} table for 1250 matt Munsell color chips 
with Munsell notation values, XYZ, xyY, RGB, CIE Lab, CIE Luv and spectral 
data.}

\source{
Spectral Color Research group, 1989
University of Kuopio, Finland
School of Computing and the Department of Physics and Mathematics
Database - Munsell Colors Matt (AOTF)
\url{http://cs.joensuu.fi/~spectral/databases/download/munsell_aotf.htm}

Parkkinen, J. P. S., Hallikainen, J. and Jaaskelainen, 1989
"Characteristic spectra of Munsell colors,"
Journal of the Optical Society of America
Vol. 6, No. 2, February 1989, pp. 318-322.
}
\references{
Spectral Color Research group, 1989
University of Kuopio, Finland
School of Computing and the Department of Physics and Mathematics
Database - Munsell Colors Matt (AOTF)
\url{http://cs.joensuu.fi/~spectral/databases/download/munsell_aotf.htm}

Parkkinen, J. P. S., Hallikainen, J. and Jaaskelainen, 1989
"Characteristic spectra of Munsell colors,"
Journal of the Optical Society of America
Vol. 6, No. 2, February 1989, pp. 318-322.
}
\author{Jose Gama}
\examples{
\dontrun{
data(MunsellSpectral)
MunsellSpectral
}
}
\keyword{datasets}

