\name{MunsellV2relativeLuminanceY}
\alias{MunsellV2relativeLuminanceY}
\title{Munsell value V to relative luminance Y}
\description{\code{MunsellV2relativeLuminanceY} Munsell value V to relative luminance Y.
}
\usage{MunsellV2relativeLuminanceY(V) }
\arguments{
  \item{V}{ Munsell value}
}
\value{
  CIE XYZ "Y"
}
\source{
Mark D. Fairchild, 2013
Color Appearance Models, 3rd Ed. Wiley-IS&T
}
\references{
Mark D. Fairchild, 2013
Color Appearance Models, 3rd Ed. Wiley-IS&T
}
\author{Jose Gama}
\examples{
MunsellV2relativeLuminanceY(5)
}
\keyword{datasets}

