\name{MunsellV2Y}
\alias{MunsellV2Y}
\title{Munsell value to CIE XYZ "Y" }
\description{\code{MunsellV2Y} Munsell value to CIE XYZ "Y" .
}
\usage{MunsellV2Y(V) }
\arguments{
  \item{V}{ Munsell value}
}
\value{
  CIE XYZ "Y"
}
\source{
ASTM, 2008
ASTM Standard D1535-08
}
\references{
ASTM, 2008
ASTM Standard D1535-08
}
\author{Jose Gama}
\examples{
MunsellV2Y(5)
}
\keyword{datasets}

