\name{NickersonColorDifference}
\alias{NickersonColorDifference}
\title{Nickerson Color Difference}
\description{\code{NickersonColorDifference} Nickerson's Color Difference.
}
\usage{NickersonColorDifference(MunsellHVC1, MunsellHVC2) }
\arguments{
  \item{MunsellHVC1}{ Munsell HVC 1}
  \item{MunsellHVC2}{ Munsell HVC 2}
}
\value{
  Delta E
}
\source{
Bruce Justin Lindbloom, 2013 
Color Calculator
\url{www.brucelindbloom.com}
}
\references{
Bruce Justin Lindbloom, 2013 
Color Calculator
\url{www.brucelindbloom.com}
}
\author{Jose Gama}
\examples{
NickersonColorDifference('10B 5/6','5B 5/4')
}
\keyword{datasets}

