\name{PreucilAngle}
\alias{PreucilAngle}
\title{Preucil Angle}
\description{\code{PreucilAngle} Preucil Angle.
}
\usage{PreucilAngle(RGBmatrix) }
\arguments{
  \item{RGBmatrix}{ RGB coordinates}
}
\value{
  Angle
}
\source{
Robert William Gainer Hunt, 1987
The Reproduction of Colour
Fountain Press	
Edition	4, illustrated
}
\references{
Robert William Gainer Hunt, 1987
The Reproduction of Colour
Fountain Press	
Edition	4, illustrated
}
\author{Jose Gama}
\examples{
PreucilAngle(c(24,72,44))
}
\keyword{datasets}

