\name{PreucilPercentGreyness}
\alias{PreucilPercentGreyness}
\title{Preucil Percentage of Greyness}
\description{\code{PreucilPercentGreyness} Preucil Percentage of Greyness.
}
\usage{PreucilPercentGreyness(RGBmatrix) }
\arguments{
  \item{RGBmatrix}{ RGB coordinates}
}
\value{
  Percentage of Greyness
}
\source{
Robert William Gainer Hunt, 1987
The Reproduction of Colour
Fountain Press	
Edition	4, illustrated
}
\references{
Robert William Gainer Hunt, 1987
The Reproduction of Colour
Fountain Press	
Edition	4, illustrated
}
\author{Jose Gama}
\examples{
PreucilPercentGreyness(c(24,72,44))
}
\keyword{datasets}

