\name{PreucilPercentHueError}
\alias{PreucilPercentHueError}
\title{Preucil Percentage of Greyness}
\description{\code{PreucilPercentHueError} Preucil Percentage of Hue Error.
}
\usage{PreucilPercentHueError(RGBmatrix) }
\arguments{
  \item{RGBmatrix}{ RGB coordinates}
}
\value{
  Percentage of HueError
}
\source{
Robert William Gainer Hunt, 1987
The Reproduction of Colour
Fountain Press	
Edition	4, illustrated
}
\references{
Robert William Gainer Hunt, 1987
The Reproduction of Colour
Fountain Press	
Edition	4, illustrated
}
\author{Jose Gama}
\examples{
PreucilPercentHueError(c(24,72,44))
}
\keyword{datasets}

