\name{RGB2CMY}
\alias{RGB2CMY}
\title{Convert sRGB coordinates to CMY}
\description{\code{RGB2CMY} Converts sRGB coordinates to CMY.
}
\usage{RGB2CMY(RGBmatrix)}
\arguments{
  \item{RGBmatrix}{ sRGB coordinates}
}
\value{
  CMY coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
RGB2CMY(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

