\name{RGB2HSL}
\alias{RGB2HSL}
\title{Convert RGB coordinates to HSL}
\description{\code{RGB2HSL} Converts RGB coordinates to HSL.
}
\usage{RGB2HSL(RGBmatrix) }
\arguments{
  \item{RGBmatrix}{ RGB coordinates}
}
\value{
  HSL coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
RGB<-c(124,63,78)
RGB2HSL(RGB)
}
\keyword{datasets}

