\name{RGB2HSV}
\alias{RGB2HSV}
\title{Convert RGB coordinates to HSV}
\description{\code{RGB2HSV} Converts RGB coordinates to HSV.
}
\usage{RGB2HSV(RGBmatrix) }
\arguments{
  \item{RGBmatrix}{ RGB coordinates}
}
\value{
  HSV coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
RGB<-c(124,63,78)
RGB2HSV(RGB)
}
\keyword{datasets}

