\name{RGB2hue}
\alias{RGB2hue}
\title{Convert sRGB coordinates to hue}
\description{\code{RGB2hue} Converts sRGB coordinates to hue.
}
\usage{RGB2hue(RGBmatrix)}
\arguments{
  \item{RGBmatrix}{ sRGB coordinates}
}
\value{
  hue
}
\author{Jose Gama}
\examples{
RGB2hue(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

