\name{RGB2LEF}
\alias{RGB2LEF}
\title{Convert RGB coordinates to LEF}
\description{\code{RGB2LEF} Converts RGB coordinates to LEF.
}
\usage{RGB2LEF(RGBmatrix) }
\arguments{
  \item{RGBmatrix}{ RGB coordinates}
}
\value{
  LEF coordinates
}
\source{
Kang, Henry R, 2006
Computational color technology
Spie Press Bellingham
}
\references{
Kang, Henry R, 2006
Computational color technology
Spie Press Bellingham
}
\author{Jose Gama}
\examples{
RGB<-c(124,63,78)
RGB2LEF(RGB)
}
\keyword{datasets}

