\name{saturationCIECAM02}
\alias{saturationCIECAM02}
\title{saturation CIECAM 2002}
\description{\code{saturationCIECAM02} saturation CIECAM 2002.
}
\usage{saturationCIECAM02(M, Q) }
\arguments{
  \item{M}{ colorfulness}
  \item{Q}{ brightness}
}
\value{
  saturation
}
\source{
Color by Wikipedians
\url{http://www.easyrgb.com/}
}
\references{
Color by Wikipedians
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}

\keyword{datasets}

