\name{saturationCIELABEvaLubbe}
\alias{saturationCIELABEvaLubbe}
\title{CIELAB saturation (Eva Lubbe)}
\description{\code{saturationCIELABEvaLubbe} CIELAB saturation (chroma 
normalized by lightness).
}
\usage{saturationCIELABEvaLubbe(L,a,b) }
\arguments{
  \item{L}{ CIELAB L}
  \item{a}{ CIELAB a}
  \item{b}{ CIELAB b}
}
\value{
  saturation
}
\source{
Color by Wikipedians
\url{http://www.easyrgb.com/}
}
\references{
Color by Wikipedians
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
saturationCIELABEvaLubbe(34.78467, 28.15159, 3.024663)
}
\keyword{datasets}

