\name{saturationCIELUV}
\alias{saturationCIELUV}
\alias{saturationCIELAB}
\title{CIELUV/CIELAB saturation}
\description{\code{saturationCIELUV} CIELUV/CIELAB saturation.
}
\usage{saturationCIELUV(u, v, un, vn) }
\arguments{
  \item{u}{ CIELAB u}
  \item{v}{ CIELAB v}
  \item{un}{ CIELAB u neutral}
  \item{vn}{ CIELAB v neutral}
}
\value{
  saturation
}
\source{
Color by Wikipedians
\url{http://www.easyrgb.com/}
}
\references{
Color by Wikipedians
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}

\keyword{datasets}

