\name{SmithPokorny2degConeFundamentals1975}
\alias{SmithPokorny2degConeFundamentals1975}
\title{Smith & Pokorny (1975) 2-deg cone fundamentals}
\usage{SmithPokorny2degConeFundamentals1975}
\description{\code{SmithPokorny2degConeFundamentals1975} 
Smith & Pokorny (1975) 2-deg cone fundamentals based on the 
CIE Judd-Vos 2-deg CMFs.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{L2}{ L-cone spectral sensitivity, L2(lambda)}
    \item{M2}{ M-cone spectral sensitivity, M2(lambda)}
    \item{S2}{ S-cone spectral sensitivity,  S2(lambda)}
  }
}
\source{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org 
}
\references{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\author{Jose Gama}
\examples{
data(SmithPokorny2degConeFundamentals1975)
SmithPokorny2degConeFundamentals1975
}
\keyword{datasets}

