\name{sprague}
\alias{sprague}
\title{Interpolates an n by w matrix of spectra, sprague}
\description{\code{sprague} Interpolates an n by w matrix of spectra, sprague.
}
\usage{sprague(spectra, f) }
\arguments{
  \item{spectra}{ spectral data}
  \item{f}{ range of wavelenghts}
}
\value{
  Interpolated spectral data
}
\source{
Stephen Westland, 2014
\url{http://www.mathworks.com/matlabcentral/fileexchange/40640-computational-colour-science-using-matlab-2e/content/sprague.m}
}
\references{
Stephen Westland, 2014
\url{http://www.mathworks.com/matlabcentral/fileexchange/40640-computational-colour-science-using-matlab-2e/content/sprague.m}
}
\author{Jose Gama}

\keyword{datasets}

