\name{StearnsStearnscorrection}
\alias{StearnsStearnscorrection}
\title{Stearns and Stearns correction}
\description{\code{StearnsStearnscorrection} Stearns and Stearns correction.
}
\usage{StearnsStearnscorrection(P) }
\arguments{
  \item{P}{ XYZ coordinates}
}
\value{
  RGB coordinates
}
\source{
Stephen Westland and Caterina Ripamonti, 2004
Computational Colour Science using MATLAB
John Wiley & Sons Ltd, pp.35
}
\references{
Stephen Westland and Caterina Ripamonti, 2004
Computational Colour Science using MATLAB
John Wiley & Sons Ltd, pp.35
}
\author{Jose Gama}

\keyword{datasets}

