\name{Stensby68.Whiteness}
\alias{Stensby68.Whiteness}
\title{Stensby Whiteness}
\description{\code{Stensby68.Whiteness} formula was developed by Mr. P. Stensby
(formerly employee of J.R. Geigy AG in US.)}
\usage{Stensby68.Whiteness(LabHunterMatrix)}
\arguments{
    \item{LabHunterMatrix}{ Lab Hunter values for illuminant C}
}
\source{
Xrite, 2012
Color iQC and Color iMatch Color Calculations Guide
Version 8.0
30 July 2012 Revision 1.0
}
\references{
Xrite, 2012
Color iQC and Color iMatch Color Calculations Guide
Version 8.0
30 July 2012 Revision 1.0
}
\author{Jose Gama}
\examples{
Stensby68.Whiteness(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

