\name{StockmanMacLeodJohnson10degConeFundamentals1993}
\alias{StockmanMacLeodJohnson10degConeFundamentals1993}
\title{Stockman & Sharpe (2000) 10-deg cone fundamentals}
\usage{StockmanMacLeodJohnson10degConeFundamentals1993}
\description{\code{StockmanMacLeodJohnson10degConeFundamentals1993} Stockman,
 MacLeod & Johnson (1993) 2-deg cone fundamentals based on the 
CIE 10-deg CMFs (adjusted to 2-deg).}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{L10}{ L-cone spectral sensitivity, L10(lambda)}
    \item{M10}{ M-cone spectral sensitivity, M10(lambda)}
    \item{S10}{ S-cone spectral sensitivity,  S10(lambda)}
  }
}
\source{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\references{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\author{Jose Gama}
\examples{
data(StockmanMacLeodJohnson10degConeFundamentals1993)
StockmanMacLeodJohnson10degConeFundamentals1993
}
\keyword{datasets}

