\name{StockmanMacLeodJohnson2degCIEadjConeFundamentals1993}
\alias{StockmanMacLeodJohnson2degCIEadjConeFundamentals1993}
\title{Stockman, MacLeod & Johnson (1993) 2-deg cone fundamentals}
\usage{StockmanMacLeodJohnson2degCIEadjConeFundamentals1993}
\description{\code{StockmanMacLeodJohnson2degCIEadjConeFundamentals1993} 
Stockman, MacLeod & Johnson (1993) 2-deg cone fundamentals based on the 
CIE 10-deg CMFs (adjusted to 2-deg).}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{L2}{ L-cone spectral sensitivity, L2(lambda)}
    \item{M2}{ M-cone spectral sensitivity, M2(lambda)}
    \item{S2}{ S-cone spectral sensitivity,  S2(lambda)}
  }
}
\source{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\references{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\author{Jose Gama}
\examples{
data(StockmanMacLeodJohnson2degCIEadjConeFundamentals1993)
StockmanMacLeodJohnson2degCIEadjConeFundamentals1993
}
\keyword{datasets}

