\name{StockmanMacLeodJohnson2degStilesBurchConeFundamentals1993}
\alias{StockmanMacLeodJohnson2degStilesBurchConeFundamentals1993}
\title{Stockman, MacLeod & Johnson (1993) 2-deg fundamentals}
\usage{StockmanMacLeodJohnson2degStilesBurchConeFundamentals1993}
\description{\code{StockmanMacLeodJohnson2degStilesBurchConeFundamentals1993} 
Stockman, MacLeod & Johnson (1993) 2-deg fundamentals based on the 
Stiles & Burch 2-deg CMFs.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{L2}{ L-cone spectral sensitivity, L2(lambda)}
    \item{M2}{ M-cone spectral sensitivity, M2(lambda)}
    \item{S2}{ S-cone spectral sensitivity,  S2(lambda)}
  }
}
\source{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\references{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\author{Jose Gama}
\examples{
data(StockmanMacLeodJohnson2degStilesBurchConeFundamentals1993)
StockmanMacLeodJohnson2degStilesBurchConeFundamentals1993
}
\keyword{datasets}

