\name{StockmanSharpe10degCMFS2000}
\alias{StockmanSharpe10degCMFS2000}
\title{Stockman & Sharpe (2000) 10-deg cone fundamentals}
\usage{StockmanSharpe10degCMFS2000}
\description{\code{StockmanSharpe10degCMFS2000} Stockman & Sharpe (2000)
 10-deg cone fundamentals based on the Stiles & Burch 10-deg CMFs linear 5nm.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{L10}{ L-cone spectral sensitivity, L10(lambda)}
    \item{M10}{ M-cone spectral sensitivity, M10(lambda)}
    \item{S10}{ S-cone spectral sensitivity,  S10(lambda)}
  }
}
\source{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\references{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\author{Jose Gama}
\examples{
data(StockmanSharpe10degCMFS2000)
StockmanSharpe10degCMFS2000
}
\keyword{datasets}

