\name{StockmanSharpe2degCMFadj2000}
\alias{StockmanSharpe2degCMFadj2000}
\title{Stockman & Sharpe (2000) 2-deg cone fundamentals}
\usage{StockmanSharpe2degCMFadj2000}
\description{\code{StockmanSharpe2degCMFadj2000} Stockman & Sharpe (2000) 
2-deg cone fundamentals based on the Stiles & Burch 10-deg CMFs 
(adjusted to 2-deg) linear 5nm.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{L2}{ L-cone spectral sensitivity, L2(lambda)}
    \item{M2}{ M-cone spectral sensitivity, M2(lambda)}
    \item{S2}{ S-cone spectral sensitivity,  S2(lambda)}
  }
}
\source{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\references{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\author{Jose Gama}
\examples{
data(StockmanSharpe2degCMFadj2000)
StockmanSharpe2degCMFadj2000
}
\keyword{datasets}

