\name{TCSdata}
\alias{TCSdata}
\title{The spectral reflectance data of 14 color test samples for CRI}
\usage{TCSdata}
\description{\code{TCSdata} is a table with the spectral reflectance data of 14 color test samples for CRI.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wavelength}{ wavelength (nm)}
    \item{TCS1}{ spectral reflectance data for sample 1}
    \item{TCS2}{ spectral reflectance data for sample 2}
    \item{TCS3}{ spectral reflectance data for sample 3}
    \item{TCS4}{ spectral reflectance data for sample 4}
    \item{TCS5}{ spectral reflectance data for sample 5}
    \item{TCS6}{ spectral reflectance data for sample 6}
    \item{TCS7}{ spectral reflectance data for sample 7}
    \item{TCS8}{ spectral reflectance data for sample 8}
    \item{TCS9}{ spectral reflectance data for sample 9}
    \item{TCS10}{ spectral reflectance data for sample 10}
    \item{TCS11}{ spectral reflectance data for sample 11}
    \item{TCS12}{ spectral reflectance data for sample 12}
    \item{TCS13}{ spectral reflectance data for sample 13}
    \item{TCS14}{ spectral reflectance data for sample 14}
  }
}
\source{
Rensselaer Polytechnic Institute
Light Sources and Color Q & A
Appendix B: MATLAB script for calculating measures of light source color: CCT, CRI, GA, and FSI
\url{http://www.lrc.rpi.edu/programs/nlpip/lightinganswers/lightsources/appendixb1.asp}
}
\references{
Rensselaer Polytechnic Institute
Light Sources and Color Q & A
Appendix B: MATLAB script for calculating measures of light source color: CCT, CRI, GA, and FSI
\url{http://www.lrc.rpi.edu/programs/nlpip/lightinganswers/lightsources/appendixb1.asp}
}
\author{Jose Gama}
\examples{
data(TCSdata)
TCSdata
}
\keyword{datasets}

