\name{VosEstevezWalraven2degConeFundamentals1990}
\alias{VosEstevezWalraven2degConeFundamentals1990}
\title{Vos, Estévez & Walraven (1990) 2-deg cone fundamentals}
\usage{VosEstevezWalraven2degConeFundamentals1990}
\description{\code{VosEstevezWalraven2degConeFundamentals1990} 
Vos, Estévez & Walraven (1990) 2-deg fundamentals based on the 
Stiles & Burch 2-deg CMFs.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{L2}{ L-cone spectral sensitivity, L2(lambda)}
    \item{M2}{ M-cone spectral sensitivity, M2(lambda)}
    \item{S2}{ S-cone spectral sensitivity,  S2(lambda)}
  }
}
\source{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\references{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\author{Jose Gama}
\examples{
data(VosEstevezWalraven2degConeFundamentals1990)
VosEstevezWalraven2degConeFundamentals1990
}
\keyword{datasets}

