\name{VosWalraven2degConeFundamentals1971}
\alias{VosWalraven2degConeFundamentals1971}
\title{Vos & Walraven(1971) 2-deg cone fundamentals}
\usage{VosWalraven2degConeFundamentals1971}
\description{\code{VosWalraven2degConeFundamentals1971} 
Vos & Walraven(1971) 2-deg cone fundamentals based on the 
CIE Judd-Vos 2-deg CMFs.}
\format{
  This data frame contains the following data:
  \describe{
    \item{wlnm}{ wavelength (nm)}
    \item{L2}{ L-cone spectral sensitivity, L2(lambda)}
    \item{M2}{ M-cone spectral sensitivity, M2(lambda)}
    \item{S2}{ S-cone spectral sensitivity,  S2(lambda)}
  }
}
\source{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\references{
The Colour & Vision Research laboratory(CVRL)
Institute of Ophthalmology, University College London
www.cvrl.org
}
\author{Jose Gama}
\examples{
data(VosWalraven2degConeFundamentals1971)
VosWalraven2degConeFundamentals1971
}
\keyword{datasets}

