\name{whitepointsilluminants}
\alias{whitepointsilluminants}
\title{White points of standard illuminants}
\usage{whitepointsilluminants}
\description{\code{whitepointsilluminants} is a table with White points 
of standard illuminants.}
\format{
  This data frame contains the following data:
  \describe{
    \item{illuminant}{ illuminant}
    \item{description}{ description}
    \item{x2}{ x2}
    \item{y2}{ y2}
    \item{x10}{ x10}
    \item{y10}{ y10}
    \item{CCT}{ CCT}
  }
}
\source{
Wikipedia, 2014
White points of standard illuminants
\url{http://en.wikipedia.org/wiki/Standard_illuminant}
}
\references{
Wikipedia, 2014
White points of standard illuminants
\url{http://en.wikipedia.org/wiki/Standard_illuminant}
}
\author{Jose Gama}
\examples{
data(whitepointsilluminants)
whitepointsilluminants
}
\keyword{datasets}

