\name{whitepointsRGB}
\alias{whitepointsRGB}
\title{Primaries for RGB color spaces}
\usage{whitepointsRGB}
\description{\code{whitepointsRGB} is a table with primaries for RGB color 
spaces.}
\format{
  This data frame contains the following data:
  \describe{
    \item{xRed}{ Primary red x}
    \item{yRed}{ Primary red y}
    \item{xGreen}{ Primary green x}
    \item{yGreen}{ Primary green y}
    \item{xBlue}{ Primary blue x}
    \item{yBlue}{ Primary blue y}
    \item{whitepointilluminant}{ illuminant}
    \item{gamma}{ gamma}
    \item{description}{ Color space name}
  }
}
\source{
Wikipedia, 2014
RGB color space
\url{http://en.wikipedia.org/wiki/RGB_color_space}
}
\references{
Wikipedia, 2014
RGB color space
\url{http://en.wikipedia.org/wiki/RGB_color_space}
}
\author{Jose Gama}
\examples{
data(whitepointsRGB)
whitepointsRGB
}
\keyword{datasets}

