\name{wlnm2XYZ}
\alias{wlnm2XYZ}
\alias{wlnm2xyz}
\title{Approximates wavelength to CIE tristimulus XYZ}
\description{\code{wlnm2XYZ} Approximates wavelength to CIE tristimulus XYZ, 
by interpolation.

\code{wlnm2xyz} Approximates wavelength to CIE xyz, by interpolation.
}
\usage{wlnm2XYZ(wavelength) }
\arguments{
  \item{wavelength}{ wavelength}
}
\value{
  CIE XYZ
}
\author{Jose Gama}
\examples{
wlnm2XYZ(555)
}
\keyword{datasets}

