\name{xyY2XYZ}
\alias{xyY2XYZ}
\title{Convert CIE CMF to XYZ}
\description{\code{xyY2XYZ} Converts CIE CMF to XYZ.
}
\usage{xyY2XYZ(xyYmatrix) }
\arguments{
  \item{xyYmatrix}{ CIE CMFs}
}
\value{
  XYZ coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
xyY2XYZ(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

