\name{XYZMoonSpencer1945}
\alias{XYZMoonSpencer1945}
\title{Approximations from wavelengths to XYZ by Moon & Spencer}
\description{\code{XYZMoonSpencer1945} Approximations from wavelengths to XYZ 
by Moon & Spencer.
}
\usage{XYZMoonSpencer1945(wavelen) }
\arguments{
  \item{wavelen}{ wavelenght data}
}
\value{
  XYZ coordinates
}
\source{
Chris Wyman Peter-Pike Sloan Peter Shirley, 2013
Simple Analytic Approximations to the CIE XYZ Color Matching Functions
Journal of Computer Graphics Techniques Vol. 2, No. 2
}
\references{
Chris Wyman Peter-Pike Sloan Peter Shirley, 2013
Simple Analytic Approximations to the CIE XYZ Color Matching Functions
Journal of Computer Graphics Techniques Vol. 2, No. 2
}
\author{Jose Gama}
\examples{
XYZMoonSpencer1945(555)
}
\keyword{datasets}

