\name{XYZperfectreflectingdiffuser}
\alias{XYZperfectreflectingdiffuser}
\title{Perfect reflecting diffuser data}
\description{\code{XYZperfectreflectingdiffuser} table with perfect reflecting 
diffuser data.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Illuminant}{ Illuminant}
    \item{X2}{ CIE tristimulus X 2 deg observer}
    \item{Y2}{ CIE tristimulus Y 2 deg observer}
    \item{Z2}{ CIE tristimulus Z 2 deg observer}
    \item{X10}{ CIE tristimulus X 10 deg observer}
    \item{Y10}{ CIE tristimulus Y 10 deg observer}
    \item{Z10}{ CIE tristimulus Z 10 deg observer}      
  }
}
\source{
Wyszecki, G. and Stiles, W.S., 1982
Color Science: Concepts and Methods, Quantitative data and formulae.
John Wiley & Sons.
}
\references{
Wyszecki, G. and Stiles, W.S., 1982
Color Science: Concepts and Methods, Quantitative data and formulae.
John Wiley & Sons.
}
\author{Jose Gama}
\examples{
data(XYZperfectreflectingdiffuser)
str(XYZperfectreflectingdiffuser)
}
\keyword{datasets}

