\name{XYZTannenbaum1974}
\alias{XYZTannenbaum1974}
\title{Approximations from wavelengths to XYZ by Tannenbaum 1974}
\description{\code{XYZTannenbaum1974} Approximations from wavelengths
to XYZ by Tannenbaum 1974.
}
\usage{XYZTannenbaum1974(wavelen) }
\arguments{
  \item{wavelen}{ wavelenght data}
}
\value{
  XYZ coordinates
}
\source{
Chris Wyman Peter-Pike Sloan Peter Shirley, 2013
Simple Analytic Approximations to the CIE XYZ Color Matching Functions
Journal of Computer Graphics Techniques Vol. 2, No. 2
}
\references{
Chris Wyman Peter-Pike Sloan Peter Shirley, 2013
Simple Analytic Approximations to the CIE XYZ Color Matching Functions
Journal of Computer Graphics Techniques Vol. 2, No. 2
}
\author{Jose Gama}
\examples{
XYZTannenbaum1974(555)
}
\keyword{datasets}

