\name{Y2MunsellV}
\alias{Y2MunsellV}
\title{CIE XYZ "Y" to Munsell value}
\description{\code{Y2MunsellV} CIE XYZ "Y" to Munsell value.
}
\usage{Y2MunsellV(Y) }
\arguments{
  \item{Y}{ Y data}
}
\value{
  Munsell value
}
\source{
ASTM, 2008
ASTM Standard D1535-08
}
\references{
ASTM, 2008
ASTM Standard D1535-08
}
\author{Jose Gama}
\examples{
Y2MunsellV(5)
}
\keyword{datasets}

