\name{Y2MunsellVtable1D1535}
\alias{Y2MunsellVtable1D1535}
\title{CIE XYZ "Y" to Munsell value formula, based on the 
ASTM Standard D1535-08}
\description{\code{Y2MunsellVtable1D1535} NLSQ regression for obtaining 
similar results to table 1 from ASTM Standard D1535-08.
}
\usage{Y2MunsellVtable1D1535(Y) }
\arguments{
  \item{Y}{ Y data}
}
\value{
  Munsell value
}
\source{
ASTM, 2008
ASTM Standard D1535-08
}
\references{
ASTM, 2008
ASTM Standard D1535-08
}
\author{Jose Gama}
\examples{
Y2MunsellVtable1D1535(5)
}
\keyword{datasets}

