\name{YIQ2RGB}
\alias{YIQ2RGB}
\title{Convert YIQ coordinates to RGB}
\description{\code{YIQ2RGB} Converts YIQ coordinates to RGB.
}
\usage{YIQ2RGB(YIQmatrix) }
\arguments{
  \item{YIQmatrix}{ YIQ coordinates}
}
\value{
  RGB coordinates
}
\source{
Alex Izvorski, Copyright 2003-2005 (Portions Copyright 2001-2003 by Alfred Reibenschuh)
Graphics/ColorObject version 0.5.0
\url{http://www.poynton.com/notes/colour_and_gamma/ColorFAQ.txt}
}
\references{
Alex Izvorski, Copyright 2003-2005 (Portions Copyright 2001-2003 by Alfred Reibenschuh)
Graphics/ColorObject version 0.5.0
\url{http://www.poynton.com/notes/colour_and_gamma/ColorFAQ.txt}
}
\author{Jose Gama}
\examples{
YIQ2RGB(c(82.949, 31.51965, 17.58261))
}
\keyword{datasets}

