\name{YUV2RGB}
\alias{YUV2RGB}
\title{Convert YUV coordinates to RGB}
\description{\code{YUV2RGB} Converts YUV coordinates to RGB.
}
\usage{YUV2RGB(YUVmatrix) }
\arguments{
  \item{YUVmatrix}{ YUV coordinates}
}
\value{
  RGB coordinates
}


\author{Jose Gama}
\examples{
YUV2RGB(c(164.898, -5.584651, 58.53939))
}
\keyword{datasets}

