\name{CheckColorLookup}
\alias{CheckColorLookup}
\title{Check that the color block number is correct}
\description{\code{CheckColorLookup} Checks that the color block number is correct.
}
\usage{CheckColorLookup(DataISCCNBS) }
\arguments{
  \item{DataISCCNBS}{ data.frame with columns MunsellSpec and Number}
}
\value{
  logic
}
\author{Glenn Davis}
\keyword{datasets}

