\name{footcandle2candela.steradian.sqmeter}
\alias{footcandle2candela.steradian.sqmeter}
\title{converts foot candle to candela steradian / square meter }
\description{\code{footcandle2candela.steradian.sqmeter} converts foot candle to candela steradian / square meter [cd*sr/m^2].
}
\usage{footcandle2candela.steradian.sqmeter(ftcl) }
\arguments{
  \item{ftcl}{ foot candle}
}
\value{
  watts / square centimeter
}
\source{
Translators cafe
\url{http://www.translatorscafe.com/cafe/EN/units-converter/illumination}
}
\references{
Translators cafe
\url{http://www.translatorscafe.com/cafe/EN/units-converter/illumination}
}
\author{Jose Gama}
\examples{
footcandle2candela.steradian.sqmeter(5)
}
\keyword{datasets}

