\name{footcandle2lux}
\alias{footcandle2lux}
\title{convert foot candle to Lumens/lux}
\description{\code{footcandle2lux} converts foot candle to Lumens/lux.
}
\usage{footcandle2lux(ftcl) }
\arguments{
  \item{ftcl}{ foot candle}
}
\value{
  Lumens/lux
}
\source{
Translators cafe
\url{http://www.translatorscafe.com/cafe/EN/units-converter/illumination}
}
\references{
Translators cafe
\url{http://www.translatorscafe.com/cafe/EN/units-converter/illumination}
}
\author{Jose Gama}
\examples{
footcandle2lux(5)
}
\keyword{datasets}

