\name{footcandle2watt.sqcentimeter}
\alias{footcandle2watt.sqcentimeter}
\title{converts foot candle to watts / square centimeter }
\description{\code{footcandle2watt.sqcentimeter} converts foot candle to watts / square centimeter [w/cm^2] (at 555 nm) .
}
\usage{footcandle2watt.sqcentimeter(ftcl) }
\arguments{
  \item{ftcl}{ foot candle}
}
\value{
  watts / square centimeter
}
\source{
Translators cafe
\url{http://www.translatorscafe.com/cafe/EN/units-converter/illumination}
}
\references{
Translators cafe
\url{http://www.translatorscafe.com/cafe/EN/units-converter/illumination}
}
\author{Jose Gama}
\examples{
footcandle2watt.sqcentimeter(5)
}
\keyword{datasets}

