\name{YUV2RGB}
\alias{YUV2RGB}
\title{Convert YUV coordinates to RGB}
\description{\code{YUV2RGB} Converts YUV coordinates to RGB.
}
\usage{YUV2RGB(YUVmatrix) }
\arguments{
  \item{YUVmatrix}{ YUV coordinates}
}
\value{
  RGB coordinates
}
\source{
Madk, Sophie Kirschner, 2014
Color space conversion 
\url{http://www.blitzbasic.com/codearcs/codearcs.php?code=2953}
}
\references{
Madk, Sophie Kirschner, 2014
Color space conversion 
\url{http://www.blitzbasic.com/codearcs/codearcs.php?code=2953}
}
\author{Jose Gama}
\examples{
YUV2RGB(c(164.898, -5.584651, 58.53939))
}
\keyword{datasets}

